+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
-## v0.0.1 (2024-09-11)
+## v0.0.6 (January 15, 2025)
 
-- New Feature: project setup
+- The first version of the GL CoreAI is released on the Google Store
+- OpenAI api model only support
+
+## v0.0.7 (January 20, 2025)
+
+- Add DeepSeek api model support